Refactor core to leverage wireit
