+++ internals/upscaler-cli/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@types/commander": "^2.12.2",
     "@types/fast-levenshtein": "^0.0.2",
+    "@internals/common": "workspace:*",
     "chalk": "^5.3.0",
     "commander": "^11.0.0",
     "fast-levenshtein": "^3.0.0",
