+++ package.json
@@ -46,6 +46,7 @@
   "devDependencies": {
     "@babel/cli": "7.22.10",
     "@babel/core": "7.22.10",
+    "@internals/common": "workspace:*",
     "@babel/plugin-proposal-export-namespace-from": "7.18.9",
     "@babel/plugin-transform-modules-commonjs": "7.22.5",
     "@babel/preset-env": "7.22.10",
