+++ packages/upscalerjs/package.json
@@ -46,29 +46,127 @@
   ],
   "wireit": {
     "build": {
-      "command": "pnpm build:only && pnpm validate:build",
+      "command": "pnpm run build:node-gpu && pnpm run build:node && pnpm run build:browser",
       "dependencies": [
         "../core:build",
         "../../models/default-model:build"
+      ],
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.esm.json",
+        "tsconfig.cjs.json",
+        "tsconfig.umd.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/**"
       ]
     },
-    "build:only": {
-      "command": "pnpm build:node-gpu && pnpm build:node && pnpm build:browser"
-    },
-    "build:browser": {
-      "command": "pnpm --filter @upscalerjs/scripts build:upscaler browser -o esm -o umd"
-    },
     "build:browser:esm": {
-      "command": "pnpm --filter @upscalerjs/scripts build:upscaler browser -o esm"
+      "command": "tsc -p ./tsconfig.esm.json --outDir ./dist/browser/esm",
+      "dependencies": [
+        "scaffold:dependencies:browser"
+      ],
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.node.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.esm.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/browser/esm/**"
+      ]
     },
     "build:browser:umd": {
-      "command": "pnpm --filter @upscalerjs/scripts build:upscaler browser -o umd"
+      "command": "pnpm run build:browser:umd:tsc && pnpm run build:browser:umd:rollup && pnpm run build:browser:umd:uglify && rimraf ./dist/browser/umd-tmp",
+      "dependencies": [
+        "scaffold:dependencies:browser"
+      ],
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.node.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.umd.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/browser/umd/**"
+      ]
+    },
+    "build:browser:umd:tsc": {
+      "command": "tsc -p ./tsconfig.umd.json --outDir ./dist/browser/umd-tmp"
+    },
+    "build:browser:umd:rollup": {
+      "command": "rollup -c ./rollup.config.mjs --input ./dist/browser/umd-tmp/umd.js --file ./dist/browser/umd/upscaler.js --name Upscaler --format umd"
+    },
+    "build:browser:umd:uglify": {
+      "command": "uglifyjs ./dist/browser/umd/upscaler.js --output ./dist/browser/umd/upscaler.min.js --compress --mangle --source-map"
+    },
+    "build:browser": {
+      "command": "pnpm run build:browser:esm && pnpm run build:browser:umd",
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.node.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.esm.json",
+        "tsconfig.umd.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/browser/**"
+      ]
     },
     "build:node": {
-      "command": "pnpm --filter @upscalerjs/scripts build:upscaler node -o cjs"
+      "command": "tsc -p ./tsconfig.cjs.json --outDir ./dist/node/cjs",
+      "dependencies": [
+        "scaffold:dependencies:node"
+      ],
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.browser.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.cjs.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/node/**"
+      ]
     },
     "build:node-gpu": {
-      "command": "pnpm --filter @upscalerjs/scripts build:upscaler node-gpu -o cjs"
+      "command": "tsc -p ./tsconfig.cjs.json --outDir ./dist/node-gpu/cjs",
+      "dependencies": [
+        "scaffold:dependencies:node"
+      ],
+      "files": [
+        "src/**/*.ts",
+        "!src/**/*.browser.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.generated.ts",
+        "package.json",
+        "vite.config.ts",
+        "tsconfig.cjs.json",
+        "tsconfig.json"
+      ],
+      "output": [
+        "dist/node-gpu/**"
+      ]
     },
     "scaffold:dependencies:browser": {
       "command": "ts-node ../../scripts/package-scripts/scaffold-dependencies.ts --src packages/upscalerjs --config packages/upscalerjs/scaffolder.ts browser"
@@ -83,7 +181,7 @@
       "command": "pnpm lint"
     },
     "prepublishOnly": {
-      "command": "pnpm lint && pnpm test && pnpm build && pnpm validate:build"
+      "command": "pnpm lint && pnpm test && pnpm build"
     },
     "format": {
       "command": "echo 'Formatting code' && prettier --write \"src/**/*.ts\""
@@ -125,13 +223,13 @@
       ]
     },
     "test:unit:browser:vite": {
-      "command": "vitest --config vitest.browser.ts",
+      "command": "vitest --config vite.browser.ts",
       "dependencies": [
         "scaffold:dependencies:browser"
       ]
     },
     "test:unit:node": {
-      "command": "vitest --config vitest.node.ts",
+      "command": "vitest --config vite.node.ts",
       "dependencies": [
         "scaffold:dependencies:node"
       ]
@@ -142,10 +240,12 @@
   },
   "scripts": {
     "build": "wireit",
-    "build:only": "wireit",
-    "build:browser": "wireit",
     "build:browser:esm": "wireit",
     "build:browser:umd": "wireit",
+    "build:browser:umd:tsc": "wireit",
+    "build:browser:umd:rollup": "wireit",
+    "build:browser:umd:uglify": "wireit",
+    "build:browser": "wireit",
     "build:node": "wireit",
     "build:node-gpu": "wireit",
     "scaffold:dependencies:browser": "wireit",
