+++ packages/upscalerjs/vite.browser.ts
